feat: kubeconfig without need for aws-iam authenticator
The kubeconfig reconcilation has been changed in 2 ways. The
kubeconfig expected by CAPI has been changed to use a token that is
based on a presign URL. This token is only valid for a maximum of
15 minutes so on every sync it updates the kubeconfig with a new token
and this means we check that the --sync-period isn't above 10 mins if
EKS is enabled.

We also generate an additional kubeconfig that a use can use that will
utilize aws-iam-authenticator or the AWS cli to get the token.